+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 ## Changelog
 
-### Version 1.0.1 (15-Oct-2025)
+### Version 1.1.0 (16-Oct-2025)
+> :warning: **Breaking Change** :warning:
+>
+> For users of version `1.0.0`, please first uninstall version `1.0.0` **before** installing version `1.1.0` (or later) via:
+> 
+> `resetConfiguration="${4:-"Uninstall"}"`
+>
+> Please feel free to reach out to the [Mac Admins Slack](https://www.macadmins.org/) (free, registration required) [#ddm-os-reminders](https://slack.com/app_redirect?channel=C09LVE2NVML) channel for assistance.
+> 
+> _Sorry for any Dan-induced headaches._
+
 - Refactored `infobuttonaction` to disable blurscreen (Pull Request #2; thanks, @TechTrekkie!)
 - Updated `message` variable to clarify update instructions
 - Tweaked `updateScriptLog` function to satisfy my CDO (i.e., the alphabetical version of "OCD")
+- Added `checkUserFocusDisplayAssertions` function to avoid interrupting users with Focus modes or Display Sleep Assertions enabled (thanks, @TechTrekkie!)
 
 ### Version 1.0.0 (14-Oct-2025)
 - First "official" release (thanks for the testing and feedback, @TechTrekkie!)